# src/validator/swagger.yaml
openapi: 3.0.0
info:
  title: Welcome to REST API DOCS with
  version: 1.0.0
  description: This is documentation on REST API for Ruth's portfolio with Typescript,node,express,mongoDb and documented by using swagger-jsdoc and swagger-ui-express
  contact:
    name: Ruth Uwamahoro
    url: https://github.com/Ruthuwamahoro/backend-side
    email: ruthuwamahoro250@gmail.com
schemes:
  - https
  - http
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key Authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer Authentication
security:
  - bearerAuth: []
  - ApiKeyAuth: []
paths:
...
  "/project/postproject":
    post:
      summary: publish an project
      description: Only authenticated user can publish project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: title
                description:
                  type: string
                  example: content
                demo:
                  type: string
                  example: description
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    type: string
                    example: project created

  # "/project/getsinglepost/{id}":
  #   get:
  #     summary: publish an article
  #     description: Only authenticated user can post blog.
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       "200":
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status:
  #                   type: string
  #                   example: ok
  #                 data:
  #                   type: object
  #                   properties:
  #                     _id:
  #                       type: string
  #                       example: id
  #                     title:
  #                       type: string
  #                       example: title
  #                     content:
  #                       type: string
  #                       example: content
  #                     description:
  #                       type: string
  #                       example: description
  #                     image:
  #                       type: string
  #                     created_at:
  #                       type: string

  # delete single project
  "/project/deleteproject/{id}":
    delete:
      summary: delete an project
      description: Only authenticated user can delete project.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    type: string
                    example: project deleted

  # update single project
  "/project/updateproject/{id}":
    patch:
      summary: update  project
      description: Only authenticated user can update project.
      consumes:
        - multipart/form-data
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                title:
                  type: string
                  example: title
                description:
                  type: string
                  example: content
                demo:
                  type: string
                  example: description
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  projects:
                    type: array
                    properties:
                      _id:
                        type: string
                        example: id
                      title:
                        type: string
                        example: title
                      description:
                        type: string
                        example: content
                      demo:
                        type: string
                        example: description